Refactor extension function declarations
